% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiextract}
\alias{multiextract}
\title{Extract seed traits from SID}
\usage{
multiextract(queryresult)
}
\arguments{
\item{queryresult}{A result from the function sidspecurl or sidspecurls}
}
\description{
This function checks a sidspecurls result and returns traits, if the queried
species has them listed in the database.
}

