% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sidspecurls}
\alias{sidspecurls}
\title{Check URLs for binomial names in SID}
\usage{
sidspecurls(sciname, sepa = " ", single = FALSE)
}
\arguments{
\item{sciname}{A vector of binomial names}

\item{sepa}{Character separating genus and species names. Defaults to " "}

\item{single}{Should the resulting data frame have only one row for each name queried? Defaults to FALSE}
}
\description{
This function checks if Seed Information Database has entries for a given
vector of binomial names, and returns the URLs for any existing entries.
}

