% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{sidtraits}
\alias{sidtraits}
\alias{sidtraits-package}
\title{sidtraits accesses the Royal Botanic Gardens Kew Seed Information Database (SID, data.kew.org/sid/).
It's a work in progress.
SID's servers are not the fastest, and have been down several times lately. If the servers are down,
sidtraits' functions will fail.
Currently scraped seed traits: 1000 seed weight (g)}
\description{
sidtraits accesses the Royal Botanic Gardens Kew Seed Information Database (SID, data.kew.org/sid/).
It's a work in progress.
SID's servers are not the fastest, and have been down several times lately. If the servers are down,
sidtraits' functions will fail.
Currently scraped seed traits: 1000 seed weight (g)
}

