% Generated by roxygen2 (4.0.0): do not edit by hand
\name{traitextract}
\alias{traitextract}
\title{Extract seed traits from SID}
\usage{
traitextract(queryresult)
}
\arguments{
\item{queryresult}{A single row result from the function sidspecurl or sidspecurls}
}
\description{
This function checks a sidspecurl result and returns traits, if the queried
species has them listed in the database. Intended for use with multiextract.
}

