% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sidseeds}
\alias{sidseeds}
\title{Extract seed traits from SID using binomial names}
\usage{
sidseeds(sciname, sepa = " ", single = FALSE)
}
\arguments{
\item{sciname}{A vector of one or more binomial names}

\item{sepa}{Character separating genus and species names. Defaults to " "}

\item{single}{Should the result be restricted to one species / queried name? Defaults to FALSE}
}
\description{
This function queries the Royal Botanic Gardens Kew Seed Information Database (SID, http://data.kew.org/sid/)
for seed traits using plant species binomial names. Currently the function returns recognised names, average
1000 seed masses and principal dispersal agents.
}
\examples{
sidseeds("Betula pendula")
sidseeds("Betula pendula", single = TRUE)
sidseeds("Betula_pendula", sepa = "_")
sidseeds(c("Betula pendula", "Anemone nemorosa"))
}

